made changes ask in PR
