+++ build.gradle
@@ -1,9 +1,9 @@
-group 'com.nitishm.userlist'
+group 'org.webonise.javacollectionassignment'
 version '1.0-SNAPSHOT'
 
 apply plugin: 'java'
 
-sourceCompatibility = 1.5
+sourceCompatibility = 1.8
 
 repositories {
     mavenCentral()
