+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'userlist'
+rootProject.name = 'javacollectionassignment'
