+++ src/main/java/org/webonise/javacollectionassignment/UserList.java
@@ -1,4 +1,4 @@
-package org.nitishm.javacollectionassignment;
+package org.webonise.javacollectionassignment;
 
 import java.util.Collections;
 import java.util.List;
